pytest test refactoring
